Move various utilities to a new KLInternal module
